# perf(core): use opcall() directly